+++ examples/02_samples_reads_assembly_bins.yaml
@@ -2,7 +2,7 @@
 # ABOUT: Coverage is derived from one unsorted .bam file.
 # ABOUT: Taxonomy is derived from `gtdb_to_ncbi_majority_vote.py` output
 # USAGE: navigate to the <examples> directory
-# USAGE: submg submit --config 02_samples_reads_assembly_bins.yaml --staging_dir <path/to/staging/dir> --logging_dir <path/to/logging/dir> --submit_samples --submit_reads --submit_assembly --submit_bins
+# USAGE: submg-cli submit --config 02_samples_reads_assembly_bins.yaml --staging_dir <path/to/staging/dir> --logging_dir <path/to/logging/dir> --submit_samples --submit_reads --submit_assembly --submit_bins
   
 STUDY: "PRJEB71644"                                                                                   # The accession of your study (which has to already exist in ENA) >>EXAMPLE: "PRJEB71644"
 METAGENOME_SCIENTIFIC_NAME: 'ant fungus garden metagenome'                                            # Taxonomic identifier of the metagenome. Check the ENA metagenome taxonomy tree to find a taxonomy ID and species name fitting your sample >>EXAMPLE: "biogas fermenter metagenome"
