+++ examples/04_reads_assembly_bins_mags.yaml
@@ -3,7 +3,7 @@
 # ABOUT: Taxonomy is derived from `gtdb_to_ncbi_majority_vote.py` output and a MANUAL_TAXONOMY_FILE
 # ABOUT: Uses the locus tag 'LTAG01' 
 # USAGE: navigate to the <examples> directory
-# USAGE: submg submit --config 04_reads_assembly_bins_mags.yaml --staging_dir <path/to/staging/dir> --logging_dir <path/to/logging/dir> --submit_reads --submit_assembly --submit_bins --submit_mags
+# USAGE: submg-cli submit --config 04_reads_assembly_bins_mags.yaml --staging_dir <path/to/staging/dir> --logging_dir <path/to/logging/dir> --submit_reads --submit_assembly --submit_bins --submit_mags
 
 STUDY: "PRJEB71644"                                                           # The accession of your study (which has to already exist in ENA) >>EXAMPLE: "PRJEB71644"
 METAGENOME_SCIENTIFIC_NAME: "biogas fermenter metagenome"                     # Taxonomic identifier of the metagenome. Check the ENA metagenome taxonomy tree to find a taxonomy ID and species name fitting your sample >>EXAMPLE: "biogas fermenter metagenome"
