+++ examples/07_assembly_bins_mags.yaml
@@ -3,7 +3,7 @@
 # ABOUT: Coverage is derived from .bam files
 # ABOUT: Taxonomy is derived from `gtdb_to_ncbi_majority_vote.py` output and a MANUAL_TAXONOMY_FILE
 # USAGE: navigate to the <examples> directory
-# USAGE: submg submit --config 07_assembly_bins_mags.yaml --staging_dir <path/to/staging/dir> --logging_dir <path/to/logging/dir> --submit_assembly --submit_bins --submit_mags
+# USAGE: submg-cli submit --config 07_assembly_bins_mags.yaml --staging_dir <path/to/staging/dir> --logging_dir <path/to/logging/dir> --submit_assembly --submit_bins --submit_mags
 
 
 STUDY: "PRJEB71644"                                                           # The accession of your study (which has to already exist in ENA) >>EXAMPLE: "PRJEB71644"
