--- requirements.txt
@@ -1,2 +1,0 @@
-# Requirements defined in setup.py
-.
