+++ setup.py
@@ -6,7 +6,7 @@ from setuptools.command.install import install
 class CustomInstall(install):
     def run(self):
         install.run(self)
-        print("\nIMPORTANT: Please run the 'submg download_webin' command.")
+        print("\nIMPORTANT: Please run the 'submg-cli download_webin' command.")
 
 setup(
     name='submg',
@@ -15,25 +15,27 @@ setup(
     install_requires=[
         'pysam>=0.19.1; sys_platform != "Windows"',
         'PyYAML>=5.4.1',
-        'Requests>=2.31.0',
+        'requests>=2.31.0',
         'tqdm>=4.64.1',
         'yaspin>=2.0',
-    ],
-    scripts=[
-        'submg/main.py',
+        'customtkinter>=5.2.2',
     ],
     include_package_data=True,
     package_data={
-        'submg': ['*.yaml','*.fa','*.fasta','*.fasta.fai','*.bam','*.bam.bai','*.tsv','*.txt']
+        'submg.resources': ['*.png'],
+        'submg': ['*.yaml', '*.fa', '*.fasta', '*.fasta.fai', '*.bam', '*.bam.bai', '*.tsv', '*.txt']
     },
     entry_points={
         'console_scripts': [
-            'submg=submg.main:main'
+            'submg=submg.main:main',
+            'submg-cli=submg.cli_main:main',
+            'submg-gui=submg.gui_main:main',
         ],
     },
     cmdclass={
         'install': CustomInstall,
     },
+
     # Metadata
     author='ttubb',
     author_email='t.tubbesing@uni-bielefeld.de',
