+++ submg/__init__.py
@@ -1,14 +1,25 @@
 # submg/__init__.py
 
-from . import statConf
-from . import configGen
-from . import utility
-from . import preflight
-from . import sampleSubmission
-from . import readSubmission
-from . import assemblySubmission
-from . import binSubmission
-from . import webinWrapper
-from . import loggingC
-from . import webinDownload
-from . import taxQuery
+from submg.core import (
+    init_argparse,
+    download_webin,
+    makecfg,
+    submit,
+    # Add other functions if needed
+)
+
+# Version information
+__version__ = '1.0.0'
+
+#from . import statConf
+#from . import configGen
+#from . import utility
+#from . import preflight
+#from . import sampleSubmission
+#from . import readSubmission
+#from . import assemblySubmission
+#from . import binSubmission
+#from . import webinWrapper
+#from . import loggingC
+#from . import webinDownload
+#from . import taxQuery
