+++ submg/gui/__init__.py
@@ -1,0 +1,3 @@
+# submg/gui/__init__.py
+
+from . import menu
