+++ submg/gui_main.py
@@ -1,0 +1,10 @@
+from .gui import menu
+
+
+def main():
+    menu.main()
+
+
+if __name__ == "__main__":
+    main()
+
