+++ submg/modules/__init__.py
@@ -1,0 +1,14 @@
+# submg/modules/__init__.py
+
+from . import statConf
+from . import configGen
+from . import utility
+from . import preflight
+from . import sampleSubmission
+from . import readSubmission
+from . import assemblySubmission
+from . import binSubmission
+from . import webinWrapper
+from . import loggingC
+from . import webinDownload
+from . import taxQuery
