+++ submg/modules/sampleSubmission.py
@@ -1,8 +1,8 @@
 import os
 
-from submg import loggingC, utility
-from submg.utility import from_config, stamped_from_config
-from submg.statConf import staticConfig
+from . import loggingC, utility
+from .utility import from_config, stamped_from_config
+from .statConf import staticConfig
 
 import requests
